fix: amountForの戻り値修正

+++ 01-refactoring/customer-2.java
@@ -45,7 +45,7 @@ class Customer {
     return result;
   }
 
-  private int amountFor(Rental each) {
+  private double amountFor(Rental each) {
     int thisAmount = 0
     switch (each.getMovie().getPriceCode()) {
       case Movie.REGULAR:
